feat(kqnodes): documentación, material de defensa y análisis de resultados
Consolida la entrega de KQNodes y el análisis experimental.

Algoritmo y pruebas:
- Elimina la matriz de costos T (código muerto) en kqnodes.py.
- Corrige y amplía la suite de tests (5/5 en verde).
- Ajustes menores en GeoMIP y QNodes (geometric, ncube, system, manager, q_nodes).

Documentación KQNodes:
- Amplía manual técnico y de usuario (k=2..k, relación con QNodes, ejemplos reales).
- Migra los diagramas de Mermaid a PlantUML (fuentes .puml + PNG).
- Genera manuales en Word, diapositivas (PPTX) y teleprompter para el video.

Tooling y análisis:
- scripts/ para volcar resultados a Excel (QNodes/GeoMIP/KQNodes) con timeout y
  reintento de casos en TIMEOUT.
- Carga eficiente de TPM (pandas float32) y límite de workers por N para evitar
  MemoryError en redes grandes (N>=22).
- scripts/analizar_resultados.py: hoja "Análisis 10-15-20" en DatosPruebas2026_1.xlsx
  con tablas resumen, 4 gráficos nativos e interpretación (phi vs k, consistencia
  QNodes vs Geometric en k=2, escalabilidad temporal).

Infra:
- .gitignore ignora las muestras TPM grandes (N20A/N22A/N25A).

+++ GeoMIP/src/Method2_Dynamic_Programming_Reformulation/src/models/core/system.py
@@ -34,9 +34,11 @@ class System:
             NCube(
                 indice=i,
                 dims=np.array(range(n_nodes), dtype=np.int8),
-                data=tpm[:, i].reshape((2,) * n_nodes)
-                if notacion == Notation.LIL_ENDIAN.value
-                else tpm[:, i][reindexar(tpm[COLS_IDX])].reshape((2,) * n_nodes),
+                data=(
+                    tpm[:, i].reshape((2,) * n_nodes)
+                    if notacion == Notation.LIL_ENDIAN.value
+                    else tpm[:, i][reindexar(tpm[COLS_IDX])].reshape((2,) * n_nodes)
+                ).astype(np.float32, copy=False),
             )
             for i in range(n_nodes)
         )
